Revert ActivityParty.Size to long[]

+++ src/Wumpus.Net.Core/Entities/Presences/Activity.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using Voltaic;
 using Voltaic.Serialization;
-using Voltaic.Serialization.Json;
 
 namespace Wumpus.Entities
 {
@@ -56,7 +55,7 @@ namespace Wumpus.Entities
         public Optional<Utf8String> Id { get; set; }
         /// <summary> used to show the party's current and maximum size </summary>
         [ModelProperty("size")]
-        public Optional<byte[]> Size { get; set; }
+        public Optional<long[]> Size { get; set; }
     }
 
     /// <summary> https://discordapp.com/developers/docs/topics/gateway#activity-object-activity-assets </summary>
